Moved css into styles
